+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 
 * [BC] PHP minimum version is now 5.5
 * [BC] Now require Aws SDK v3
+* [BC] To avoid name clashes, module name has been renamed from `Aws` to `AwsModule`.
 
 ## 1.2.0